Create Game
I've been trying to fix a problem with this code. when i type in the terminal "east2", instead of making a text appear it just leaves a blank space twice. then the text "Not an option" appears, even though it is an option. If i type east2 one last time then the text appears, but not before everything else. Is there any way to fix that?